+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/client/CTIClient.java
@@ -37,6 +37,7 @@ import java.util.Map;
 import java.util.concurrent.ExecutionException;
 
 import com.wazuh.contentmanager.model.ctiapi.ConsumerInfo;
+import com.wazuh.contentmanager.model.ctiapi.ContextChanges;
 import com.wazuh.contentmanager.settings.PluginSettings;
 
 /**
@@ -50,10 +51,11 @@ public class CTIClient extends HttpClient {
 
     private static final Logger log = LogManager.getLogger(CTIClient.class);
 
-    private static final String API_BASE_URL = PluginSettings.getInstance().getCtiBaseUrl();
     private static final String CONSUMER_INFO_ENDPOINT =
-            "/catalog/contexts/vd_1.0.0/consumers/vd_4.8.0";
-    private static final String CONSUMER_CHANGES_ENDPOINT = "/changes";
+            "/catalog/contexts/" + PluginSettings.CONTEXT_ID + "/consumers/" + PluginSettings.CONSUMER_ID;
+    private static final String CONSUMER_CHANGES_ENDPOINT = CONSUMER_INFO_ENDPOINT + "/changes";
+
+    private static CTIClient INSTANCE;
 
     /** Enum representing the query parameters used in CTI API requests. */
     public enum QueryParameters {
@@ -85,12 +87,7 @@ public class CTIClient extends HttpClient {
      * URL.
      */
     private CTIClient() {
-        super(URI.create(API_BASE_URL));
-    }
-
-    /** Singleton holder pattern ensures lazy initialization in a thread-safe manner. */
-    private static class CTIClientHolder {
-        private static final CTIClient INSTANCE = new CTIClient();
+        super(URI.create(PluginSettings.getInstance().getCtiBaseUrl()));
     }
 
     /**
@@ -98,8 +95,11 @@ public class CTIClient extends HttpClient {
      *
      * @return The singleton instance of {@code CTIClient}.
      */
-    public static CTIClient getInstance() {
-        return CTIClientHolder.INSTANCE;
+    public static synchronized CTIClient getInstance() {
+        if (INSTANCE == null) {
+            INSTANCE = new CTIClient();
+        }
+        return INSTANCE;
     }
 
     /**
@@ -108,17 +108,37 @@ public class CTIClient extends HttpClient {
      * @param fromOffset The starting offset (inclusive) for fetching changes.
      * @param toOffset The ending offset (exclusive) for fetching changes.
      * @param withEmpties A flag indicating whether to include empty values (Optional).
-     * @return A {@link SimpleHttpResponse} containing the API response.
+     * @return {@link ContextChanges} instance with the current changes.
      */
-    public SimpleHttpResponse getChanges(String fromOffset, String toOffset, String withEmpties) {
+    public ContextChanges getChanges(String fromOffset, String toOffset, String withEmpties) {
+        XContent xContent = XContentType.JSON.xContent();
         Map<String, String> params = contextQueryParameters(fromOffset, toOffset, withEmpties);
-        return sendRequest(Method.GET, CONSUMER_CHANGES_ENDPOINT, null, params, (Header) null);
+        SimpleHttpResponse response =
+                sendRequest(Method.GET, CONSUMER_CHANGES_ENDPOINT, null, params, (Header) null);
+        if (response == null) {
+            log.error("No response from CTI API Changes endpoint");
+            return null;
+        }
+        if (response.getCode() != HttpStatus.SC_OK) {
+            log.error("CTI API Changes endpoint returned an error: {}", response.getBody());
+        }
+        log.debug("CTI API Changes endpoint replied with status: [{}]", response.getCode());
+        try {
+            return ContextChanges.parse(
+                    xContent.createParser(
+                            NamedXContentRegistry.EMPTY,
+                            DeprecationHandler.THROW_UNSUPPORTED_OPERATION,
+                            response.getBodyBytes()));
+        } catch (IOException | IllegalArgumentException e) {
+            log.error("Failed to fetch changes information", e);
+        }
+        return null;
     }
 
     /**
      * Fetches the entire CTI catalog from the API.
      *
-     * @return A {@link SimpleHttpResponse} containing the API response with the catalog data.
+     * @return A {@link ConsumerInfo} object containing the catalog information.
      */
     public ConsumerInfo getCatalog() {
         XContent xContent = XContentType.JSON.xContent();
@@ -127,6 +147,9 @@ public class CTIClient extends HttpClient {
         if (response == null) {
             log.error("No response from CTI API");
             return null;
+        }
+        if (response.getCode() != HttpStatus.SC_OK) {
+            log.error("CTI API returned an error: {}", response.getBody());
         }
         log.debug("CTI API replied with status: [{}]", response.getCode());
         try {
@@ -203,6 +226,7 @@ public class CTIClient extends HttpClient {
         } catch (ExecutionException e) {
             log.error("Snapshot download failed: {}", e.getMessage());
         } catch (InterruptedException e) {
+            Thread.currentThread().interrupt(); // Restore the interrupted status
             log.error("Snapshot download was interrupted: {}", e.getMessage());
         }
     }