+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/model/ctiapi/ConsumerInfo.java
@@ -151,7 +151,16 @@ public class ConsumerInfo implements ToXContentObject {
      * @return Context name as a String
      */
     public String getContext() {
-        return context;
+        return this.context;
+    }
+
+    /**
+     * Getter for the last offset number
+     *
+     * @return Last offset number as a long
+     */
+    public Long getLastOffset() {
+        return this.lastOffset;
     }
 
     /**
@@ -160,6 +169,6 @@ public class ConsumerInfo implements ToXContentObject {
      * @return A Snapshot URL
      */
     public String getLastSnapshotLink() {
-        return lastSnapshotLink;
+        return this.lastSnapshotLink;
     }
 }